+++ lib/tarantula/src/core/tarantula.Chrome.scala
@@ -45,7 +45,7 @@ import strategies.throwUnsafely
 
 object Chrome extends Browser(t"chrome"):
   def launch(port: Int)(using WorkingDirectory, Monitor): Server logs ExecEvent =
-    val server: Process["chromedriver", Text] = sh"chromedriver --port=$port".fork()
+    val server: Job["chromedriver", Text] = sh"chromedriver --port=$port".fork()
     sleep(100L)
     Server(port, server)
